feat: integrate asynchronous user creation with transaction support
- Added `CreateUserTx` method to `Store` interface and implemented it in `SQLStore`.
- Moved `TransferTx` related code to a new file `tx_transfer.go`.
- Updated `rpc_create_user.go` to use `CreateUserTx` for user creation and email verification task distribution.
- Added mock methods for `CreateUserTx` in `mock/store.go`.
- Introduced `tx_create_user.go` to handle user creation transactions with rollback/commit logic.
